pyproject.toml
[project]
name = "llm-policy-faithfulness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "pyyaml",
    "python-dotenv",
]
